requirements.txt
# Core ML and Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Steam API
requests>=2.31.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework (for deployment)
flask>=3.0.0
flask-cors>=4.0.0

# Jupyter for development
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional utilities
python-dotenv>=1.0.0
